fix(cli): cast x86_64 gregs index to usize on Linux
libc::REG_RSP is an i32 constant on Linux x86_64, but uc_mcontext.gregs
indexes by usize. Cast fixes the Linux/Ubuntu CI build.

+++ cli/src/iced.rs
@@ -1019,7 +1019,7 @@ mod unix {
         }
         unsafe {
             let uc = &*(ucontext as *const libc::ucontext_t);
-            uc.uc_mcontext.gregs[libc::REG_RSP] as usize
+            uc.uc_mcontext.gregs[libc::REG_RSP as usize] as usize
         }
     }
